# Improved performance of non0hist