@Datseris 2018-12-04
yes

_yes_ **yeees** !!! Nice, thank you very much for this! Can I ask two questions for demographic purposes?

1. where did you find out about this issue?
2. Do you also use **DynamicalSystems.jl** ?