@oneill 2018-12-08
>     1. where did you find out about this issue?
>     2. Do you also use **DynamicalSystems.jl** ?

I found this issue on Bountysource.
I did not know about **DynamicalSystems.jl** before that, but I enjoyed working on this issue :thumbsup: