@Datseris 2018-12-08
@oneill thanks very much for this! Don't forget to request the bounty!